analysis_service/requirements.txt
kafka-python==2.0.2
awkward==2.5.1
numpy==1.26.2
scipy==1.11.4
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
